fix: capture expected and observed failure evidence
